--- Neo.Storage.RocksDB/RocksDbPlugin.cs
@@ -1,36 +1,0 @@
-﻿using Neo.Persistence;
-
-namespace Neo.Plugins.Storage
-{
-    public class RocksDbPlugin : Plugin, IStoragePlugin
-    {
-        public readonly RocksDBStore _store;
-
-        public override string Name => nameof(RocksDbPlugin);
-
-        /// <summary>
-        /// Constructor
-        /// </summary>
-        public RocksDbPlugin()
-        {
-            Settings.Load(GetConfiguration());
-            _store = new RocksDBStore(Settings.Default.Path);
-        }
-
-        /// <summary>
-        /// Configure
-        /// </summary>
-        public override void Configure()
-        {
-            // Can't configure the path because NeoSystem store in cache some values
-
-            Settings.Load(GetConfiguration());
-        }
-
-        /// <summary>
-        /// Get store
-        /// </summary>
-        /// <returns>RocksDbStore</returns>
-        public IStore GetStore() => _store;
-    }
-}
